% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_states_unemp_heatmap.R
\name{create_states_unemp_heatmap}
\alias{create_states_unemp_heatmap}
\title{Title}
\usage{
create_states_unemp_heatmap(fred_key = fredKey)
}
\arguments{
\item{fred_key}{}
}
\description{
Title
}
