% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unemp_heatmap.R
\name{create_unemp_heatmap}
\alias{create_unemp_heatmap}
\title{Creates a Heatmap of Unemployment in the United States in the style of a stock market board}
\usage{
create_unemp_heatmap(
  fred_key = fredKey,
  color_up = "#d40509",
  color_down = "#046438",
  nas = "black",
  background = "black",
  borders = "black"
)
}
\arguments{
\item{fred_key}{A FRED API Key}

\item{color_up}{A color for when unemployment increases (and when it is unchanged)}

\item{color_down}{A color for when unemployment decreases}

\item{nas}{Color for NA cells}

\item{background}{Table background color}

\item{borders}{Border color}
}
\value{
A PNG Table
}
\description{
Creates a Heatmap of Unemployment in the United States in the style of a stock market board
}
\examples{
create_unemp_heatmap()
}
