% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_api_credentials.R
\name{enter_api_credentials}
\alias{enter_api_credentials}
\title{A script to load or save API keys for the MacroTools package. If the keys are already saved, it loads them into the environment. If they are not loaded, it produces a message to insert them.
#############################################################
NOTE: After saving an API Key, YOU MUST RESTART THE R SESSION
#############################################################}
\usage{
enter_api_credentials(fred_api_key = NULL, bls_api_key = NULL)
}
\arguments{
\item{fred_key}{An API key from FRED}

\item{bls_key}{An API key from BLS}
}
\value{
It creates the API keys in the global environment. Maybe not the best practice
}
\description{
A script to load or save API keys for the MacroTools package. If the keys are already saved, it loads them into the environment. If they are not loaded, it produces a message to insert them.
#############################################################
NOTE: After saving an API Key, YOU MUST RESTART THE R SESSION
#############################################################
}
\examples{
enter_api_credentials()  # Loads already saved API Keys into the environment
enter_api_credentials(fred_api_key = "FRED_API_KEY_GOES_HERE")  # Saves or updates the FRED KEY
enter_api_credentials(bls_api_key = "BLS_API_KEY_GOES_HERE")  # Saves or updates the BLS KEY
}
