% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bea_table.R
\name{get_bea_nipa}
\alias{get_bea_nipa}
\title{Download NIPA table from BEA website}
\usage{
get_bea_nipa(TableName = NULL, Frequency = "Q", Year = "ALL", UserID = NULL)
}
\arguments{
\item{TableName}{A string with table name (see https://apps.bea.gov/api/_pdf/bea_web_service_api_user_guide.pdf)}

\item{Frequency}{A string with frequency (e.g. 'A,Q') - no spaces}

\item{Year}{A string with year (e.g. '2002,2003'); no spaces - default 'ALL'}

\item{UserID}{A string with BEA API id}
}
\value{
A JSON list with data and metadata
}
\description{
Download NIPA table from BEA website
}
\examples{
\dontrun{my_id <- options()$BEA_API
#' get_bea_nipa("T10106", "Q", "2022,2023", my_id)}
}
