% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bls_refweek.R
\name{get_bls_refweek}
\alias{get_bls_refweek}
\title{Obtain BLS reference week}
\usage{
get_bls_refweek(startyear, endyear)
}
\arguments{
\item{startyear}{A character or integer scalar}

\item{endyear}{A character or integer scalar}
}
\value{
A vector of intervals representing the BLS reference week
}
\description{
Obtain BLS reference week
}
\examples{
get_bls_refweek(2020, 2022)
get_bls_refweek("2020", "2022")
}
