% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bls_series.R
\name{get_bls_series}
\alias{get_bls_series}
\title{Download BLS series}
\usage{
get_bls_series(idlist, startyear = NULL, endyear = NULL, api_key = NULL)
}
\arguments{
\item{idlist}{A list of series ID}

\item{startyear}{An integer indicating the start year (if NULL, use endyear - 4)}

\item{endyear}{An integer indicating the end year (if NULL, use current year)}

\item{api_key}{A character string of the API key obtained from the BLS website}
}
\value{
A list of two tibbles: data and catalog
}
\description{
Download BLS series
}
\examples{
\dontrun{my_id <- options()$blskey
#' get_bls_series(list("CES0000000001"), api_key = my_id)}
}
