% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_claims.R
\name{get_claims}
\alias{get_claims}
\title{Downloads unemployment claims data from FRED (weekly or monthly)}
\usage{
get_claims(
  start_year = NULL,
  end_year = NULL,
  monthly = FALSE,
  latest = FALSE,
  fred_key = fredKey
)
}
\arguments{
\item{start_year}{First year of data to get. Default is five years from 'end_year'}

\item{end_year}{Last year of data to get. Default is today's year}

\item{monthly}{If FALSE (default), returns weekly claims. If TRUE, returns monthly averages}

\item{latest}{Gets the latest complete observation}

\item{fred_key}{A FRED API key}
}
\value{
An xts object with unemployment claims data
}
\description{
Downloads unemployment claims data from FRED (weekly or monthly)
}
\examples{
claims_weekly <- get_claims()
claims_monthly <- get_claims(monthly = TRUE)
claims_latest <- get_claims(latest = TRUE)
}
