% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_BLS.R
\name{get_from_BLS}
\alias{get_from_BLS}
\title{Downloads unemployment data from Fred}
\usage{
get_from_BLS(years = 5, seriesIDs, format = "long", BLS_key = blsKey)
}
\arguments{
\item{years}{Number of years of data to get}

\item{BLS_key}{A BLS KEy}

\item{geography}{Whether to include all states}

\item{demography}{Wheter to include race/ethnicity}

\item{latest}{Gets the latest complete unemployment figures}

\item{fred_key}{A FRED API}
}
\value{
An xts object with unemployment data
}
\description{
Downloads unemployment data from Fred
}
\examples{
unemployment <- get_unemployment()
}
