% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDP.R
\name{get_GDP}
\alias{get_GDP}
\title{Downloads Real GDP from Fred for the US, all US States, or a selection of states.}
\usage{
get_GDP(
  start_year = NULL,
  end_year = NULL,
  geography = "National",
  latest = FALSE,
  fred_key = fredKey
)
}
\arguments{
\item{start_year}{First year of data to get. Default is five years from 'end_year'}

\item{end_year}{Last year of data to get. Default is today's year}

\item{geography}{'National' (default), 'State', or list of states like c("DC","NM")}

\item{latest}{Gets the latest complete GDP figures}

\item{fred_key}{A FRED API}
}
\value{
An xts object with GDP data
}
\description{
Downloads Real GDP from Fred for the US, all US States, or a selection of states.
}
\examples{
real_GDP <- get_GDP()
real_GDP_from_2020 <- get_GDP(start_year = 2020)
states_real_GDP <- get_GDP(geography = c("DC","NC"), latest = TRUE)
}
