% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inflation.R
\name{get_inflation}
\alias{get_inflation}
\title{Downloads unemployment data from Fred}
\usage{
get_inflation(
  years = 5,
  monthly = FALSE,
  main_variables = FALSE,
  fred_key = fredKey
)
}
\arguments{
\item{years}{Number of years of data to get}

\item{geography}{Whether to include all states}

\item{demography}{Wheter to include race/ethnicity}
}
\value{
An xts object with unemployment data
}
\description{
Downloads unemployment data from Fred
}
\examples{
unemployment <- get_unemployment()
}
