% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jolts.R
\name{get_jolts}
\alias{get_jolts}
\title{Downloads monthly JOLTS data from FRED}
\usage{
get_jolts(
  start_year = NULL,
  end_year = NULL,
  latest = FALSE,
  fred_key = fredKey
)
}
\arguments{
\item{start_year}{First year of data to get. Default is five years from 'end_year'}

\item{end_year}{Last year of data to get. Default is today's year}

\item{latest}{Gets the latest complete JOLTS figures}

\item{fred_key}{A FRED API key}
}
\value{
An xts object with JOLTS data
}
\description{
Downloads monthly JOLTS data from FRED
}
\examples{
jolts <- get_jolts()
jolts_latest <- get_jolts(latest = TRUE)
jolts_2010_2020 <- get_jolts(start_year = 2010, end_year = 2020)
}
