% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_indeces.R
\name{get_price_indeces}
\alias{get_price_indeces}
\title{Downloads price index data from FRED and the BLS.}
\usage{
get_price_indeces(
  start_year = NULL,
  end_year = NULL,
  seasonally_adjusted = TRUE,
  fred_key = fredKey,
  BLS_key = blsKey
)
}
\arguments{
\item{start_year}{First year of data to get. Default is five years from 'end_year'}

\item{end_year}{Last year of data to get. Default is today's year}

\item{seasonally_adjusted}{If TRUE (default), returns seasonally adjusted series. If FALSE, returns not seasonally adjusted series.}

\item{fred_key}{A FRED API key}

\item{BLS_key}{A BLS API key}
}
\value{
An xts object with price index data
}
\description{
Variables included:
CPI, Core CPI, PCE, Core PCE, Producer Price Index,
Food, Food at home, Food away, Energy, Energy commodities,
Gasoline, Energy services, Electricity, Utility gas service,
Commodities less food and energy, New vehicles, Used cars and trucks,
Apparel, Medical care commodities, Services less energy,
Shelter, Transportation, Medical care services, Fuel Oil
}
\examples{
price_indeces <- get_price_indeces()
price_indeces_nsa <- get_price_indeces(seasonally_adjusted = FALSE)
}
