% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unemployment.R
\name{get_unemployment}
\alias{get_unemployment}
\title{Downloads monthly unemployment data from Fred and the BLS}
\usage{
get_unemployment(
  start_year = NULL,
  end_year = NULL,
  geography = "National",
  demography = FALSE,
  latest = FALSE,
  fred_key = fredKey,
  BLS_key = blsKey
)
}
\arguments{
\item{start_year}{First year of data to get. Default is five years from 'end_year'}

\item{end_year}{Last year of data to get. Default is today's year}

\item{geography}{'National' (default), 'State', or list of states like c("DC","NM")}

\item{demography}{Include national levels of unemployment by gender/race/ethnicity}

\item{latest}{Gets the latest complete unemployment figures}

\item{fred_key}{A FRED API}

\item{BLS_key}{A BLS KEy}
}
\value{
An xts object with unemployment data
}
\description{
Downloads monthly unemployment data from Fred and the BLS
}
\examples{
unemployment <- get_unemployment()
unemployment_state_demography <- get_unemployment(geography = c("DC","NC"), demography = TRUE)
unemployment_latest <- get_unemployment(latest = TRUE)
unemployment_2020_2021 <- get_unemployment(start_year = 2020, end_year = 2021)
}
