% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JECTools-package.R, R/MacroTools-package.R
\docType{package}
\name{MacroTools-package}
\alias{MacroTools-package}
\title{MacroTools: A Compendium of Tools and Packages for Macroecononomic Analysis}
\description{
This is a collection of tools and packages to conduct macroeconomic analysis using United States data. It relies heavily on other packages. To work properly, two API Keys most be provided before executing any function: fredKey <- "FRED_API_KEY_GOES_HERE" and blsKey <- "BLS_API_KEY_GOES_HERE".

This is a collection of tools and packages to conduct macroeconomic analysis using United States data. It relies heavily on other packages. To work properly, two API Keys most be provided before executing any function: fredKey <- "FRED_API_KEY_GOES_HERE" and blsKey <- "BLS_API_KEY_GOES_HERE".
}
\author{
\strong{Maintainer}: Romero Alfredo \email{aaromero@ncat.edu} (\href{https://orcid.org/0000-0001-7241-0824}{ORCID})

}
\keyword{internal}
