% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nberrecession}
\alias{nberrecession}
\title{NBER Recession Dates}
\format{
\subsection{\code{nberrecession}}{

A tibble with 2 columns:
\describe{
\item{peak}{Start of a recession}
\item{trough}{End of a recession}
}
}
}
\source{
\url{http://data.nber.org/data/cycles/}
}
\usage{
nberrecession
}
\description{
Us Business Cycle Expansions and Contractions
Contractions (recessions) start at the peak of a business cycle
and ent at the trough.
}
\details{
Data as of July 19, 2023
}
\keyword{datasets}
