% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_tibble.R
\name{transform_to_tibble}
\alias{transform_to_tibble}
\title{It transforms an xts object to a tibble}
\usage{
transform_to_tibble(data = xts_obj)
}
\arguments{
\item{xts_obj}{The xts object to transform}
}
\value{
A tibble
}
\description{
It transforms an xts object to a tibble
}
\examples{
data_tsibble <- transform_to_tsibble()
}
