% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_states_unemp_heatmap.R
\name{create_states_unemp_heatmap}
\alias{create_states_unemp_heatmap}
\title{A script to create a heatmap of the United States Unemployment Rate}
\usage{
create_states_unemp_heatmap(fred_key = fredKey)
}
\arguments{
\item{fred_key}{}
}
\value{
A PNG
}
\description{
A script to create a heatmap of the United States Unemployment Rate
}
\examples{
create_states_unemp_heatmap
}
