% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_api_credentials.R
\name{enter_api_credentials}
\alias{enter_api_credentials}
\title{A script to enter API keys for the JECTools package}
\usage{
enter_api_credentials(fred_api_key = NULL, bls_api_key = NULL)
}
\arguments{
\item{fred_key}{An API key from FRED}

\item{bls_key}{An API key from BLS}
}
\value{
It creates the API keys in the global environment. Maybe not the best practice
}
\description{
A script to enter API keys for the JECTools package
}
\examples{
enter_api_credentials()
}
