% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_BLS.R
\name{get_from_BLS}
\alias{get_from_BLS}
\title{Downloads unemployment data from Fred}
\usage{
get_from_BLS(
  start_year = NULL,
  end_year = NULL,
  seriesIDs,
  format = "long",
  BLS_key = blsKey
)
}
\arguments{
\item{start_year}{First year of data to get. Default is five years from 'end_year'}

\item{end_year}{Last year of data to get. Default is today's year}

\item{seriesIDs}{Vector of codes from the BLS to get.}

\item{format}{Output format to get: Default is 'long', but it can return 'wide' and 'individual' for individual series}

\item{BLS_key}{A BLS Key}
}
\value{
An xts object
}
\description{
Downloads unemployment data from Fred
}
\examples{
get_from_BLS(seriesIDs = "LNS14000000")
get_from_BLS(seriesIDs = c("LNS14000000","CES0000000001"))
get_from_BLS(end_year = 2019, start_year = 2012, seriesIDs = "LNS14000000")
get_from_BLS(seriesIDs = c("LNS14000000","CES0000000001"), format = 'individual')
get_from_BLS(seriesIDs = c("LNS14000000","BDU0000000000000000110001LQ5"), format = 'wide')
}
