% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDP.R
\name{get_GDP}
\alias{get_GDP}
\title{Downloads unemployment data from Fred}
\usage{
get_GDP(
  years = 5,
  geography = "National",
  perCapita = FALSE,
  latest = FALSE,
  fred_key = fredKey
)
}
\arguments{
\item{years}{Number of years of data to get}

\item{geography}{Whether to include all states}

\item{latest}{Gets the latest complete unemployment figures}

\item{fred_key}{A FRED API}

\item{BLS_key}{A BLS KEy}
}
\value{
An xts object with unemployment data
}
\description{
Downloads unemployment data from Fred
}
\examples{
unemployment <- get_unemployment()
}
