% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inflation.R
\name{get_inflation}
\alias{get_inflation}
\title{Computes inflation data. It uses unseasonably adjusted data for yearly computed inflation and seasonally adjusted data for monthly inflation. The main inflation variables are:
"CPIAUCSL",	      'CPI',
"CPILFESL",	      'Core CPI',
"PCEPI",	          'PCE',
"PCEPILFE",	      'Core PCE',
"PPIFIS",	        'Producer Price Index'
The additional inflation variables are:
"CPIUFDSL",	      'Food',
"CUSR0000SAF11",	  'Food at home',
"CUSR0000SEFV",	  'Food away',
"CPIENGSL",	      'Energy',
"CUSR0000SACE",	  'Energy commodities',
"CUSR0000SETB01",	'Gasoline',
"CUSR0000SEHF",	  'Energy services',
"CUSR0000SEHF01",	'Electricity',
"CUSR0000SEHF02",	'Utility gas service',
"CUSR0000SACL1E",	'Commodities less food and energy',
"CUSR0000SETA01",	'New vehicles',
"CUSR0000SETA02",	'Used cars and trucks',
"CPIAPPSL",	      'Apparel',
"CUSR0000SAM1",	  'Medical care commodities',
"CUSR0000SASLE",	  'Services less energy',
"CUSR0000SAH1",	  'Shelter',
"CUSR0000SAS4",	  'Transportation',
"CUSR0000SAM2",	  'Medical care services')}
\usage{
get_inflation(
  start_year = NULL,
  end_year = NULL,
  monthly = FALSE,
  main_variables = TRUE,
  fred_key = fredKey,
  BLS_key = blsKey
)
}
\arguments{
\item{start_year}{First year of data to get. Default is five years from 'end_year'}

\item{end_year}{Last year of data to get. Default is today's year}

\item{monthly}{Whether to calculate MoM \% Change. The default is FALSE}

\item{main_variables}{Whether to include only CPI, Core CPI, PCE, Core PCE, and Producer Price Index. The default is TRUE}
}
\value{
An xts object with inflation data
}
\description{
Computes inflation data. It uses unseasonably adjusted data for yearly computed inflation and seasonally adjusted data for monthly inflation. The main inflation variables are:
"CPIAUCSL",	      'CPI',
"CPILFESL",	      'Core CPI',
"PCEPI",	          'PCE',
"PCEPILFE",	      'Core PCE',
"PPIFIS",	        'Producer Price Index'
The additional inflation variables are:
"CPIUFDSL",	      'Food',
"CUSR0000SAF11",	  'Food at home',
"CUSR0000SEFV",	  'Food away',
"CPIENGSL",	      'Energy',
"CUSR0000SACE",	  'Energy commodities',
"CUSR0000SETB01",	'Gasoline',
"CUSR0000SEHF",	  'Energy services',
"CUSR0000SEHF01",	'Electricity',
"CUSR0000SEHF02",	'Utility gas service',
"CUSR0000SACL1E",	'Commodities less food and energy',
"CUSR0000SETA01",	'New vehicles',
"CUSR0000SETA02",	'Used cars and trucks',
"CPIAPPSL",	      'Apparel',
"CUSR0000SAM1",	  'Medical care commodities',
"CUSR0000SASLE",	  'Services less energy',
"CUSR0000SAH1",	  'Shelter',
"CUSR0000SAS4",	  'Transportation',
"CUSR0000SAM2",	  'Medical care services')
}
\examples{
inflation <- get_inflation()
inflation_since_2020 <- get_inflation(start_year = 2020)
inflationMoM <- get_inflation(monthly = TRUE, main_variables = FALSE)
}
