% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nominal_GDP.R
\name{get_nominal_GDP}
\alias{get_nominal_GDP}
\title{Downloads Real GDP from Fred for the US, all US States, or a selection of states.}
\usage{
get_nominal_GDP(
  start_year = NULL,
  end_year = NULL,
  geography = "National",
  latest = FALSE,
  fred_key = fredKey
)
}
\arguments{
\item{geography}{'National' (default), 'State', or list of states like c("DC","NM")}

\item{latest}{Gets the latest complete GDP figures}

\item{fred_key}{A FRED API}

\item{years}{Number of years of data to get}
}
\value{
An xts object with GDP data
}
\description{
Downloads Real GDP from Fred for the US, all US States, or a selection of states.
}
\examples{
nominal_GDP <- get_nominal_GDP()
states_nominal_GDP <- get_nominal_GDP(geography = c("DC","NC"), latest = TRUE)
}
