% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_indeces.R
\name{get_price_indeces}
\alias{get_price_indeces}
\title{Downloads unemployment data from Fred}
\usage{
get_price_indeces(years = 5, fred_key = fredKey, BLS_key = blsKey)
}
\arguments{
\item{years}{Number of years of data to get}

\item{fred_key}{A fred API key}

\item{BLS_key}{A bls API Key}
}
\value{
An xts object with unemployment data
}
\description{
Downloads unemployment data from Fred
}
\examples{
unemployment <- get_unemployment()
}
