% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_indeces.R
\name{get_price_indeces}
\alias{get_price_indeces}
\title{Downloads unemployment data from Fred}
\usage{
get_price_indeces(years = 5, fred_key = fredKey)
}
\arguments{
\item{years}{Number of years of data to get}

\item{geography}{Whether to include all states}

\item{demography}{Wheter to include race/ethnicity}
}
\value{
An xts object with unemployment data
}
\description{
Downloads unemployment data from Fred
}
\examples{
unemployment <- get_unemployment()
}
