% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unemployment.R
\name{get_unemployment}
\alias{get_unemployment}
\title{Downloads unemployment data from Fred}
\usage{
get_unemployment(
  years = 5,
  geography = "National",
  demography = FALSE,
  latest = FALSE,
  fred_key = fredKey,
  BLS_key = blsKey
)
}
\arguments{
\item{years}{Number of years of data to get}

\item{geography}{Whether to include all states}

\item{demography}{Wheter to include race/ethnicity}

\item{latest}{Gets the latest complete unemployment figures}

\item{fred_key}{A FRED API}

\item{BLS_key}{A BLS KEy}
}
\value{
An xts object with unemployment data
}
\description{
Downloads unemployment data from Fred
}
\examples{
unemployment <- get_unemployment()
}
