% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unemployment.R
\name{get_unemployment}
\alias{get_unemployment}
\title{Downloads monthly unemployment data from Fred and the BLS}
\usage{
get_unemployment(
  start_year = NULL,
  end_year = NULL,
  geography = "National",
  demography = FALSE,
  latest = FALSE,
  underutilization = FALSE,
  fred_key = fredKey,
  BLS_key = blsKey
)
}
\arguments{
\item{start_year}{First year of data to get. Default is five years from 'end_year'}

\item{end_year}{Last year of data to get. Default is today's year}

\item{geography}{'National' (default), 'State', or list of states like c("DC","NM")}

\item{demography}{Include national levels of unemployment by gender/race/ethnicity}

\item{latest}{Gets the latest complete unemployment figures}

\item{underutilization}{If TRUE, downloads U-1 through U-6 measures (U1RATE, U2RATE, UNRATE as U-3, U4RATE, U5RATE, U6RATE)}

\item{fred_key}{A FRED API key}

\item{BLS_key}{A BLS API key}
}
\value{
An xts object with unemployment data
}
\description{
Downloads monthly unemployment data from Fred and the BLS
}
